% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasedhmc-package.R
\docType{package}
\name{debiasedhmc-package}
\alias{debiasedhmc-package}
\alias{debiasedhmc}
\title{debiasedhmc}
\description{
Unbiased HMC estimators with couplings
}
\details{
This package contains scripts to remove the bias of HMC estimators.
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>
}
\keyword{package}
