% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{gaussian_max_coupling}
\alias{gaussian_max_coupling}
\title{Maximal coupling of two multivariate Normal distributions}
\usage{
gaussian_max_coupling(mu1, mu2, Sigma1, Sigma2)
}
\arguments{
\item{mu1}{mean of first distribution}

\item{mu2}{mean of second distribution}

\item{Sigma1}{Variance of first distribution}

\item{Sigma2}{Variance of second distribution}
}
\description{
Sample from maximal coupling of two multivariate Normal distributions,
specified through their means and covariance matrices.
}
