% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_kernel.R
\name{get_hmc_kernel}
\alias{get_hmc_kernel}
\title{Get Hamiltonian Monte Carlo kernels}
\usage{
get_hmc_kernel(logtarget, gradlogtarget, stepsize, nsteps, dimension)
}
\arguments{
\item{logtarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{gradlogtarget}{function to compute gradient of target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{stepsize}{step size in the leap-frog integrator}

\item{nsteps}{number of leap-frog steps}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled HMC kernels. These kernels can then be used in the function \code{\link{coupled_chains}}.
}
