% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvnormal.R
\name{get_mvnormal}
\alias{get_mvnormal}
\title{Get Multivariate Normal target}
\usage{
get_mvnormal(dimension, mean_target, Sigma_target)
}
\arguments{
\item{dimension}{dimension of the target}

\item{Sigma_target}{variance}

\item{mu1}{mean}
}
\value{
list with keys \code{logtarget} and \code{gradlogtarget},
ready to be used e.g. in \code{\link{get_hmc_kernel}}.
}
\description{
Defines a multivariate Normal target
with prescribed dimension, mean, variance.
}
