% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{H_bar}
\alias{H_bar}
\title{Compute unbiased estimators from coupled chains}
\usage{
H_bar(c_chains, h = function(x) x, k = 0, m = 1)
}
\description{
Compute the proposed unbiased estimators given the output
of the 'coupled_chains' function. The integral of interest is that of the function h,
which can be multivariate. The choice of k and m must be such that m is at most the choice
made when running coupled_chains, and k must be less than m.
}
