% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_c_chains.R
\name{histogram_c_chains}
\alias{histogram_c_chains}
\title{histogram_c_chains}
\usage{
histogram_c_chains(c_chains, component, k, K, breaks = NULL, nclass = 30)
}
\arguments{
\item{c_chains}{list containing pairs of chains generated by \code{\link{coupled_chains}}}

\item{component}{index of the desired marginal distribution}

\item{k}{tuning parameter representing 'burn-in'}

\item{K}{tuning parameter representing 'total number of iteration per estimator'}

\item{breaks}{vector of breaks (default to NULL)}

\item{nclass}{desired number of classes; either specify this or \code{breaks} (default to 30)}
}
\description{
Compute histograms based on coupled Markov chains.
}
