% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mala_kernel.R
\name{get_mala_kernel}
\alias{get_mala_kernel}
\title{Get adaptive coupled Metropolis adjusted Langevin algorithm (MALA) kernels
that switches between synchrous coupling and maximal coupling}
\usage{
get_mala_kernel(logtarget, gradlogtarget, stepsize, dimension, adaptive_tol)
}
\arguments{
\item{logtarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{gradlogtarget}{function to compute gradient of target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{stepsize}{standard deviation of Gaussian increment}

\item{dimension}{dimension of the target distribution}

\item{adaptive_tol}{threshold below which maximal coupling is triggered}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled MALA kernels. These kernels can then be used in the function \code{\link{coupled_chains}}.
}
