% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_c_chains.R
\name{plot_histogram}
\alias{plot_histogram}
\title{plot_histogram}
\usage{
plot_histogram(histogram, with_bar = TRUE)
}
\arguments{
\item{histogram}{list resulting from \code{\link{histogram_c_chains}}}

\item{with_bar}{boolean specifying whether to plot vertical bars for point estimates,
or only grey rectangles for confidence intervals.}
}
\description{
Plot histograms produced by \code{\link{histogram_c_chains}}.
}
