% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_couplings.R
\name{rgamma_coupled}
\alias{rgamma_coupled}
\title{Sample from maximally coupled Gamma}
\usage{
rgamma_coupled(alpha1, alpha2, beta1, beta2)
}
\arguments{
\item{alpha1}{shape parameter of first distribution}

\item{alpha2}{shape parameter of second distribution}

\item{beta1}{rate parameter of first distribution}

\item{beta2}{rate parameter of second distribution}
}
\description{
Sample from maximally coupled Gamma distributions with given parameters alpha1, alpha2, beta1, beta2,
where alpha is the shape and beta the rate, i.e. the log-density at x is

alpha * log(beta) - lgamma(alpha) + (alpha-1) * log(x) - beta x
}
