% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{rnorm_max_coupling}
\alias{rnorm_max_coupling}
\title{Maximal coupling of two univariate Normal distributions}
\usage{
rnorm_max_coupling(mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{mu1}{mean of first distribution}

\item{mu2}{mean of second distribution}

\item{sigma1}{standard deviation of first distribution}

\item{sigma2}{standard deviation of second distribution}
}
\description{
Sample from maximal coupling of two univariate Normal distributions,
specified through their means and standard deviations.
}
