% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{unbiased_estimator}
\alias{unbiased_estimator}
\title{Unbiased estimator}
\usage{
unbiased_estimator(logtarget, single_kernel, coupled_kernel, rinit,
  h = function(x) x, k = 0, m = 1, max_iterations = Inf)
}
\arguments{
\item{logtarget}{function evaluating the log target density}

\item{single_kernel}{function taking a state (in a vector), its log density and an iteration, and returning
a list with a key named \code{chain_state} containing the next state and its log density \code{current_pdf}.}

\item{coupled_kernel}{function taking two states (in two vectors), their log densities and an iteration,
and returning a list with keys \code{chain_state1}, \code{chain_state2}, \code{current_pdf1} and \code{current_pdf2}.}

\item{rinit}{function taking no arguments are returning an initial state for a Markov chain.}

\item{h}{test function (possibly vector-valued)}

\item{k}{burn-in parameter (default to 0)}

\item{m}{time average parameter (will be proportional to the computing cost if meeting occurs before \code{m},
default to 1).}

\item{max_iterations}{number of iterations at which the function stops if it is still running  (default to Inf).}
}
\description{
Sample two MCMC chains, each following \code{single_kernel} marginally,
and \code{coupled_kernel} jointly, until min(max(tau, m), max_iterations), where tau
is the first time at which the two chains meet (i.e. take the same value exactly).
Or more precisely, they meet with a delay of one, i.e. X_t = Y_{t-1}. The chains
are initialized from the distribution provided in \code{rinit}.

 See \code{\link{get_hmc_kernel}}
for an example of function returning the appropriate kernels.
}
